Revert "FILE requires including stdio"
This reverts commit acad95830201880ea84a4920a72524f722428953. Reason is
preparation for libfcgi support.

stdio.h was included in the first place for the fp parameter of
recvline() declared in cgi.h. If we want to link libfcgi we can not
include stdio.h anymore, because libfcgi replaces all i/o functions for
using your old CGI code with the fastcgi protocol.

To still use libcgi, you should include stdio.h or fcgi_stdio.h
before cgi.h to be sure FILE is a defined type.

Signed-off-by: Alexander Dahl <post@lespocky.de>