set headers_initialized in cgi_init()
With fastcgi the binary linked against libcgi does not run once and
exits immediately anymore, but in an endless loop cgi_init(), some more
functions and cgi_end() are called over and over again. Setting
headers_initialized on load and not resetting it, means on the next
cycle it was already set and no headers where sent, resulting in an
useless response to the webserver.

Signed-off-by: Alexander Dahl <post@lespocky.de>