introduce fastcgi support
FastCGI is a different way to let your web application communicate with
the webserver. It is designed to allow an easy transition from CGI based
solutions. This is done by linking libfcgi and including their headers
for I/O replacing the usual stdio.h headers.

For libcgi this is made optional by offering a CMake option. However
selecting this option and running the new binaries the old CGI way still
works.

Signed-off-by: Alexander Dahl <post@lespocky.de>